ListView Overhaul & DataStorage Modifications
Fixed a variety of minor bugs in the ListView code which was preventing
compilation. (A couple of errors in file naming and locations due to
differences between our code and the code ListView was based on)
Also began modifications to DataStorage (Arrays to ArrayLists) so that
it matches the requirements of the ListView code. (Will be finished in
the next commit)