Started on link between Spinner and ListView
The MainActivity and the ListActivity are now two separate things.
The SpinnerListener class has been removed, but a SpinnerListener is now
implemented directly in the main activty - due to the requirement of
fields in MainActivity to connect to the ListActivity.

Currently all selections on the spinner default to the Shingles page.
This will be fixed in a later update. (Also coming: more data sets!)