Data Sets and UI - Also, renamed the app.
Added the organics and glass data sets, and put them in the
spinner/list.

Also made some modifications to the MainActivity. Now, touching the
Spinner resets the selection to "Select a Type". Choosing "Select a
Type" no longer triggers a call to the ListActivity (and thus prevents a
null pointer error). The reason for the reset is it allows for the user
to select a type "twice" in a row, by adding in a buffer - the selection
the reset makes.

The app is now called "Maryland Recycling Centers" within the
application. This is subject to change.
More data sets in next commit.