Data Sets and UI update
Data Sets:
-gypsum
-toners
-metals (partially...)
-oil/antifreeze

UI:
Created a third display page that is triggered by selecting one of the
choices in the ListView. It passes on the Center object of that
position. The center's object is displayed there - and this is where the
Google Maps Interface will be.

Also, some more generic cleaning of old, unused objects and comments.