DataStorage Modification
Based off of yesterday's experience with latency issues while creating
the DataStorage object in ListActivity, the DataStorage class has been
redesigned to only instantiate and return the array that is called, when
it is called - not all of them every time. (Really should have done this
to start, but that's what you get when you don't have spare time to
think.)

Should be at least slightly faster now.